arredondando valor total

+++ projeto/model/pedido.py
@@ -5,7 +5,7 @@ class Pedido(db.Model):
     id = db.Column(db.Integer, primary_key=True,  autoincrement = True)
     nome = db.Column(db.String(100), nullable=False)
     telefone = db.Column(db.String(100), nullable=False)
-    total_pagar = db.Column(db.Integer, nullable = False)
+    total_pagar = db.Column(db.Float, nullable = False)
